+++ CHANGELOG.md
@@ -534,3 +534,6 @@
 * Updated Unity to 2020.3.43f1
 * Updated package dependencies
 * PersistState and ConfigState now preserve unknown properties when saving/loading (JsonExtensionData accessible via AdditionalData)
+
+# 4.x.y vNext
+* Fix ScriptExecuteSpecial breaking when no activator is passed